Move Checkout core function to CoreEditor + create file for shadowmapGUI + rename VSSEnabledto  VCSEnabled
