+++ com.unity.render-pipelines.core/CoreRP/Editor/Shadow.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61ef33a54404cb747aad57d04e535154
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
